# docs/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FamilyRelocation API",
    "description": "API for managing Orthodox Jewish family relocation to Union County, NJ",
    "version": "v1"
  },
  "paths": {
    "/api/applicants": {
      "get": {
        "tags": [
          "Applicants"
        ],
        "summary": "Gets a paginated list of applicants with search, filter, and sort options.",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "description": "Page number (1-based). Default: 1",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "Number of items per page. Default: 20, Max: 100",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "description": "Search term to match against name, email, or phone.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "BoardDecision",
            "in": "query",
            "description": "Filter by board decision (Pending, Approved, Rejected, Deferred).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "City",
            "in": "query",
            "description": "Filter by city.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CreatedAfter",
            "in": "query",
            "description": "Filter by created date (on or after).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "CreatedBefore",
            "in": "query",
            "description": "Filter by created date (on or before).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "description": "Sort field. Options: familyName, createdDate, boardReviewDate. Default: createdDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortOrder",
            "in": "query",
            "description": "Sort order. Options: asc, desc. Default: desc",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Applicants"
        ],
        "summary": "Creates a new applicant (family) and their housing search.\r\nPublicly accessible for board approval applications.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicantCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicantCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicantCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/applicants/{id}": {
      "get": {
        "tags": [
          "Applicants"
        ],
        "summary": "Gets an applicant by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Applicants"
        ],
        "summary": "Updates an existing applicant's basic information.\r\nCannot update: board decision, created date, applicant ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicantCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicantCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicantCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressDto": {
        "required": [
          "city",
          "state",
          "street",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "nullable": true
          },
          "street2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "fullAddress": {
            "type": "string",
            "description": "Formatted full address: \"123 Main St, Union, NJ 07083\"",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ChildDto": {
        "required": [
          "age",
          "gender"
        ],
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "school": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateApplicantCommand": {
        "required": [
          "husband"
        ],
        "type": "object",
        "properties": {
          "husband": {
            "$ref": "#/components/schemas/HusbandInfoDto"
          },
          "wife": {
            "$ref": "#/components/schemas/SpouseInfoDto"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildDto"
            },
            "nullable": true
          },
          "currentKehila": {
            "type": "string",
            "nullable": true
          },
          "shabbosShul": {
            "type": "string",
            "nullable": true
          },
          "housingPreferences": {
            "$ref": "#/components/schemas/HousingPreferencesDto"
          }
        },
        "additionalProperties": false,
        "description": "Command to create a new applicant (family) and their housing search."
      },
...
      "HousingPreferencesDto": {
        "type": "object",
        "properties": {
          "budgetAmount": {
            "type": "number",
            "description": "Maximum budget amount in USD.",
            "format": "double",
            "nullable": true
          },
          "minBedrooms": {
            "type": "integer",
            "description": "Minimum number of bedrooms required.",
            "format": "int32",
            "nullable": true
          },
          "minBathrooms": {
            "type": "number",
            "description": "Minimum number of bathrooms required.",
            "format": "double",
            "nullable": true
          },
          "requiredFeatures": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of required features (e.g., \"basement\", \"garage\", \"yard\").",
            "nullable": true
          },
          "shulProximity": {
            "$ref": "#/components/schemas/ShulProximityPreferenceDto"
          },
          "moveTimeline": {
            "type": "string",
            "description": "When the family plans to move.\r\nValid values: Immediate, ShortTerm, MediumTerm, LongTerm, Extended, Flexible, NotSure, Never",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO for housing preferences when creating an applicant."
      },
      "HusbandInfoDto": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "fatherName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhoneNumberDto"
            },
            "nullable": true
          },
          "occupation": {
            "type": "string",
            "nullable": true
          },
          "employerName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PhoneNumberDto": {
        "required": [
          "number"
        ],
        "type": "object",
        "properties": {
          "number": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ShulProximityPreferenceDto": {
        "type": "object",
        "properties": {
          "preferredShulIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "List of preferred shul IDs (optional).",
            "nullable": true
          },
          "maxWalkingDistanceMiles": {
            "type": "number",
            "description": "Maximum walking distance in miles.",
            "format": "double",
            "nullable": true
          },
          "maxWalkingTimeMinutes": {
            "type": "integer",
            "description": "Maximum walking time in minutes.",
            "format": "int32",
            "nullable": true
          },
          "anyShulAcceptable": {
            "type": "boolean",
            "description": "Whether any shul within distance is acceptable (default: true)."
          }
        },
        "additionalProperties": false,
        "description": "DTO for shul proximity preferences."
      },
      "SpouseInfoDto": {
        "required": [
          "firstName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "maidenName": {
            "type": "string",
            "nullable": true
          },
          "fatherName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhoneNumberDto"
            },
            "nullable": true
          },
          "occupation": {
            "type": "string",
            "nullable": true
          },
          "employerName": {
            "type": "string",
            "nullable": true
          },
          "highSchool": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateApplicantCommand": {
        "required": [
          "husband",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "husband": {
            "$ref": "#/components/schemas/HusbandInfoDto"
          },
          "wife": {
            "$ref": "#/components/schemas/SpouseInfoDto"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildDto"
            },
            "nullable": true
          },
          "currentKehila": {
            "type": "string",
            "nullable": true
          },
          "shabbosShul": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Command to update an existing applicant's basic information.\r\nCannot update: board decision, created date, applicant ID."
      }
    }
  },
...
}